Create longterm.html
This page will link to the main.html when clicked on link. It has two functions, inflation and interest which allow the user to see his/her money inflation over a fixed time and the interest they can gain on their principle over a fixed time. The time is inserted by the User.