Create loggin.html
This will be the first page the user sees. They will have to either login or register so that way we can verify them as a client of ours.
This page will link to the main html page if the password and username is accepted. If not then it will state it and remain on this page. 
Also features a registration button that links to that registration page.